:truck: Move `transform_derivative` into `registration.utils`
